+++ bootchart.h
@@ -100,8 +100,8 @@ extern int samples;
 extern int cpus;
 extern int len;
 extern double hz;
-extern int scale_x;
-extern int scale_y;
+extern double scale_x;
+extern double scale_y;
 extern int overrun;
 extern double interval;
